This commit makes several minor changes ....
(1) The polymer code now allows bond lengths to be set independent of segment size.
This allows for overlapping sphere models rather than simply tangent spheres.  This
change adds a new line to the input file in the interaction parameters section.

(2) This version incorporates many options for treatment of the direct correlation
functions in the polymer code.  Specifically we need to be able to mix dcfs in the
context of continuation investigations.  This changes format of the polymer input section
of the input files.

(3) Added a continuation type which is based on a prefactor multiplying a direct
correlation function.  This facilitates continuation from one dcf to another.

(4) Committing a bug fix in loca_bord.c needed to make binodal calculations work properly.

--LJDF