This version correctly calls the Schur complement solver if HAVE_DFT_SCHUR_SOLVER is defined at compile time.  The solver is still buggy, but I am working on that.
I ended up adding the fragment ", void * aux_info" (without quotes) to a number of function
calls.  This is why so many files are changed.  I needed this argument in order to get
information between the "solve_problem" function and the "newton_solver" function.
Because LOCA also calls newton_solver, I had to add this argument to a bunch of other
functions in order to get the aux_info pointer through the call levels.

I tested this change and everything seems to work fine.  I don't know if LOCA
interface I changed are part of some published interface.  If so, I may need to
use a different approach.

The added file is a cheat-script to help me compile the new solver files without
changing the autotools files for now.