----------------------------------------------------------------------  	src/dft_schur_epetra_operator.cpp src/dft_schur_solver.cpp  	src/dft_schur_solver.h  ----------------------------------------------------------------------
This version will correctly run on a single processor and use the explicit
inverse of the A11 block.  The Schur solver is enabled by using -DHAVE_DFT_SCHUR_SOLVER
in the CFLAGS and CXXFLAGS arguments to configure.  Note that this solver
only works for hard sphere models with implicit nonlocal density equations.