Added new file that contains some pseudo-code to scope out a new data structure. This proposed data structure refers to all discretized degrees-of-freedom in [node][unk#] 2d arrays and never in 1 long array. There should no longer be any need for any integer manipulations in Tramonto that look like  (inode*Nunk_per_node + iunk).
The purpose of this pseudo code is (1) to map out the "strategy"
that Laurie was asking for and (2) redefining the dft_SolutionManager
interface for Mike.

I included very abbreviated fills to demonstrate how the solution array
is accessed and how the contributions can be added to the residual (Rhs)
and Jacobian. I found the data structure and interface convenient to work
with for this very small trial.