Revised version of prototype code using new data structures resulting from conversation with Mike.
The main changes stem from the observation that  we, in
Tramonto, do not need to know about physics blocks. Those
are solution methods. We only need to say which equation
number and which unknown number the matrix entry concerns
and the solver can decide what block that corresponds to.
This makes the Tramonto code simpler and gets rid of arguments
to the dft_SolverManager methods to do with Physics Blocks.

I reworked the fill section for these changes, and put more
detail in the POISSON example. I added the density contribution
to the equation in addition to the previous LaPlace contribution.
This shows how entries in the diagonal block (derivative
of Poisson equation with respect to Poisson unknown) are
filled separately from off-diagonal blocks (derivative of
Poisson equation with respect to density unknown).

I added calls to a new method in dft_SolverManager which is adds
a single element to a Matrix instead of an array.  Here it is,
with variable types, next to the regular one.

  addOneMatrixValue(int inode, int iunk, double value, int jbox, int junk);

  addMatrixValues(int inode, int iunk, int numEntries,
                  double* values, int* jboxIndices, int junk);