Changed 2D indexing into solution vector to be x[phys#][node#] (instead of [node#][phys#]). Changed the argument order to all insert* methods in dft_SolverManager.hpp so this information pair is also in this order: (physicsID, nodeID, ...).
In reading through Mike's documentation to setRhs(double **),
I noticed that he had assumed the reverse indexing of the
2D residual and solution arrays than I had used in
the prototype file. I realized that his was the
correct indexing for making use of the MultiVector
infrastructure. With this ordering, we can now extract
a vector of density unknowns as
   double* dens = x[iDens];  //where Phys[iDens]==DENSITY;
which will make the fills and post-processing easier.