Reworked  dft_guess.c using new data structures. This file sets   xOwned[Nunk_per_node][Nnodes_per_proc] as opposed to   x[Nunk_per_node][Nnodes_box] and then using importR2c(x, xOwned) to set up the initial x.
Changes in main and newton files so that the box-sized  x[][]
is the array that is created n main, passed to the solver, and
passed to the ouput routines. I had thought this could be  xLocal
but too many of the postprocessing routines use off-processor
information, such as finite differenced flux calculations.