----------------------------------------------------------------------  Modified Files:  	configure configure.ac src/dft_fill_mf.c  ----------------------------------------------------------------------
Added -lteuchos to configure and configure.ac.  Without this,the linking
of dft executable was not working.

Add prefix "dft_" to a call of solvermanager_insertrhsvalue in dft_fill_mf.c


NOTE:  The makefiles for Tramonto have the annoying problem that .o files
are inserted into the main Tramonto/src directory.  This is bad because it
does not allow multiple simultaneous builds of Tramonto.  Also, make clean
does not work when done in a build directory other than the main Tramonto
directory.