----------------------------------------------------------------------
 Modified Files:
 	src/Makefile.in src/dft_PolyA22_Epetra_Operator.cpp
 ----------------------------------------------------------------------

Added an entry to Makefile.in for a new file Laurie added.  Added some
small logic to PolyA22 operator.  Something strange is still going on
there.  It appears the the density dependence on CMS is not a diagonal,
which is different than I assumed.  This may be the source of the differences
in results.