----------------------------------------------------------------------  Modified Files:  	src/dft_A22Matrix_Epetra_Operator.cpp  	src/dft_A22Matrix_Epetra_Operator.hpp  	src/dft_HardSphereLinProbMgr.cpp  	src/dft_HardSphereLinProbMgr.hpp  	src/dft_hardsphere_lin_prob_mgr_wrapper.cpp  	src/dft_hardsphere_lin_prob_mgr_wrapper.h src/dft_newton.c  ----------------------------------------------------------------------
Added new functionality to handle hardsphere problems where the A22
block is not diagonal.  In fact, this is now the default and you must
assert that A22 is diagonal if you want it treated that way.  There is
a new function:

int dft_hardsphere_lin_prob_mgr_seta22blockisdiagona(void * linprobmgr, int isa22diagonal);

That must be called with TRUE as the second argument if you want A22 to
be treated as a diagonal matrix.  Otherwise it will be treated as a
general matrix.  Right now the general matrix is preconditioned with
IFPACK ILU (hard-coded).  We need to set up a protocol for getting
parameter (that are different from Aztec) values to the new solvers.

Laurie, this is the feature you needed for handling A22 as a general
matrix.  Let me know how it goes.