----------------------------------------------------------------------  Modified Files:  	dft_A22Matrix_Epetra_Operator.cpp  	dft_A22Matrix_Epetra_Operator.hpp dft_BasicLinProbMgr.cpp  	dft_BasicLinProbMgr.hpp dft_HardSphereA11_Epetra_Operator.cpp  	dft_HardSphereA11_Epetra_Operator.hpp  	dft_HardSphereLinProbMgr.cpp dft_HardSphereLinProbMgr.hpp  ----------------------------------------------------------------------
Making the addition of std::map insertion approach to all other classes.
The hard sphere solver time for 2D_hs (from the paper CaseStudies) went from 85 sec
to 49 total time.  The solver time went from 61 to 25 secs.

I made the same changes to the basic solver, but have not completely confirmed
the correctness for the basic solver.  However, it compiles and I would be surprised
if it did not work.