----------------------------------------------------------------------  Modified Files:  	Fluids_DFT.tex mics-dft.aux mics-dft.bbl mics-dft.blg  	mics-dft.dvi mics-dft.log mics-dft.toc  ----------------------------------------------------------------------
This is close to a final version for turning over to Rick.  There is one paragraph that
I didn't know how to write (Extended potentials).  There is the final parallel in proposed
work that needs to discuss the merging of Fluid and QM DFTs.  I do not know what to say here yet.