----------------------------------------------------------------------  Committing in .
 Modified Files:
 	configure configure.ac
 ----------------------------------------------------------------------
Modified configure.ac to process --with-blas and --with-lapack options
like Trilinos does.  Also allows for more automatic detection of
blas and lapack.  Added FFLAGS option for the case where the value of
FLIBS might change based on a compiler option.  (I think this could be
an issue when there is a 32 or 64 bit option.)