----------------------------------------------------------------------  Committing in .
 Modified Files:
 	configure config/acx_blas.m4 config/acx_lapack.m4
 ----------------------------------------------------------------------
Tramonto also had Trilinos bug 1409.

I modified the blas and lapack search .m4 files to stop searching if a user
specified blas or lapack library cannot be found.

This only applies to blas or lapack libraries specified via --with-blas and
--with-lapack.  It is not reasonable to apply this to other options such as
--with-libs and --with-ldflags.

In the past if a user specified blas or lapack library was not found, the .m4
would continue to try to find blas or lapack in different places.
This was confusing for users because they might have been trying to use an
optimized blas library, but ended up with a different blas library.
Unless a person looked through all of the configure output, this was not
noticable.