----------------------------------------------------------------------  Committing in .
 Modified Files:
 	Makefile.in aclocal.m4 configure configure.ac
 	Examples/Makefile.in Examples/test_script src/Makefile.am
 	src/Makefile.in src/Tramonto_config.h.in
 ----------------------------------------------------------------------
I am making a few changes:

1) I added a --with-noxloca option for Andy Salinger.  It doesn't do
anything yet because Andy has to add the files to be conditionally
compiled to src/Makefile.am.

2) Updated the list of Fortran compilers to be consistent with Trilinos.
The most important addition was to add gfortran, since that is the
replacement for f77 for GCC 4.x and beyond.

3) Updated Examples/test_script.  It doesn't work completely yet, but
I am trying to get more examples working for possible use with the
test harness in the future.