Added "NOXLOCA" support, to hook Tramonto to these Trilinos packages. It is currently working with NOX but has not been expanded to use LOCA yet. The implementation is with Trilinos' development branch and should remain stable for the upcoming 7.0 release.
To compile in the capability, you mush configure with:
   --with-noxloca
(Thanks Jim!!)
Trilinos must have been configured with  --enable-nox --enable-loca

To access the capability, select in the input file:
  Continuation Method = -2

The behavior of this solver is the same as the old newton
solver for problems that do not require damping
(Tramonto's  UPDATE FRAC = 100 percent). I've selected a
backtracking algorithm in NOX as the default so the code
can also converge in cases where damping is needed, but
the convergence behavior will be different. Only time will
tell if this approach is more/less robust and more/less
efficient than our application-specific update_soln damping
algorithm.

The only place the new capability intersects with the old
code is this "if .. else" test dft_newton.c.
-------------------------------------
#ifdef HAVE_NOXLOCA
  if (Loca.method == -2)
    NOXLOCA_Solver(x, xOwned);
  else
#endif
  if (Loca.method != -1)
    iter = solve_continuation(x, x2);
  else
    iter = newton_solver(x, NULL);
-------------------------------------