----------------------------------------------------------------------  Committing in .
 Modified Files:
 	src/Makefile.am src/Makefile.in src/NOXLOCA_Solver.cpp
 	src/NOXLOCA_Tramonto_Group.hpp src/NOXLOCA_Tramonto_Vector.hpp
 	src/dft_guess.c src/dft_mesh.c src/dft_mesh_surfaces.c
 	src/dft_newton.c src/NEW/dft_guess.c src/NEW/dft_mesh.c
 	src/NEW/dft_newton.c src/NEW/guess_save.c
 Removed Files:
 	src/az_aztec.h src/az_aztec_defs.h
 ----------------------------------------------------------------------
Tramonto was still using separate versions of az_aztec.h and
az_aztec_defs.h.  The version of these files that are supplied with
Trilinos are more up to date.  After pointing to the more up to date
files and removing the old files from the repository and the build
system, I was able to remove three undef statements that Andy was
using to avoid differences between the old and new aztec files.
I also had to go through several Tramonto source files and change calls
to min and max to AZ_MIN and AZ_MAX, as these are the new names of
these functions.  min and max were probably too general and that
could have easily caused problems.

I made sure that Tramonto still compiles on my machine.