----------------------------------------------------------------------  Committing in .
 Modified Files:
 	Makefile.in aclocal.m4 configure configure.ac
 	Examples/Makefile.in src/Makefile.in
 Added Files:
 	config/tac_arg_header_check.m4
 ----------------------------------------------------------------------
Fixing a problem with the noxloca option that required us to have
-DHAVE_CONFIG_H in the list of configure arguments.  I wrote a new
macro for header checking based on one that NOX currently uses, but
I removed unnecessary complexity for our case.

For Tramonto to properly configure under the new setup when
--with-noxloca is used, it is necessary to do one of two things:

1) (Preferred) Use --with-incdirs= instead of --with-cxxflags and
--with-cflags for passing the -I that includes the path to the
installed Trilinos headers.  This is the way that Trilinos generally
gets include info.

2) In addition to --with-cxxflags and --with-cflags, the -I information
for Trilinos headers also needs to be passed to CPPFLAGS.

We could also think about moving in the direction of a --with-trilinos
option, but that would prove unnecessary if we package Tramonto with
a subset of Trilinos as we spoke about last week at SJU.