----------------------------------------------------------------------  Modified Files:  	dft_BasicLinProbMgr.cpp dft_BasicLinProbMgr.hpp  	dft_HardSphereLinProbMgr.cpp  	dft_basic_lin_prob_mgr_wrapper.cpp  	dft_basic_lin_prob_mgr_wrapper.h  ----------------------------------------------------------------------
Added a new function to dft_basic_lin_prob_mgr_wrapper and dft_BasicLinProbMgr
that allows the user to specify a list of node IDs for coarsened nodes.

  The prototype for the C function is:

  int dft_linprobmgr_setcoarsenednodeslist(void * linprobmgr, int numgids, int * gids);

The function should be called only once, prior to calling

  int dft_linprobmgr_finalizeblockstructure(void * linprobmgr);


I also added support for coarsened nodes in the HardSphereLinProbMgr class.  The equations
for these nodes will be placed at the bottom of the A22 block.  I have not verified that this
code is correct yet, since I need calls to the new C function in place first.