This commit will require everyone to slightly modify the way that they configure Tramonto.  We no longer will have to (or be able to) use the --with-indirs and --with-ldflags options for specifying the locations of Trilinos headers and libraries.  Instead, I have a created a slightly simpler --with-trilinos option that will serve both of these purposes. For example, if Trilinos is installed in /home/user/TrilinosInstall (ie --prefix=/home/user/TrilinosInstall was used as a configure option for Trilinos), then simply add
--with-trilinos=/home/user/TrilinosInstall

to your list of Tramonto configure options.

This change was made as a result of the recent configuration discussions
on the tramonto-developers mail list.  The reason that this new option isrequired and the old method will not work is because of the shift to taking
advantage of the Trilinos Makefile.export capability, which is also a part
of this commit.  This change clears up problems like the one that we werehaving with Amesos a couple months ago (if Trilinos was configured with
Amesos, Tramonto suffered a link error because of the Amesos-dependent code
that is compiled into Ifpack).  Any solution that I could think of to
support the old and new configuration options was either a bit of a hack or
defeated the purpose for using Makefile.export, so I thought it would be best
to make this change now while the user base is small, rather than try to
support the old system indefinitely.

Also included in this commit are some configure.ac changes related to theswitch to Makefile.export, as well as updates for all the test harness scripts,
sample configure scripts, and the README file.  All of these changes are in
response to the new required configure option.  I also added a couple of
perl scripts to the distribution tarball.  These scripts are used by
Makefile.export.

In addition, Tramonto will now get information about BLAS and LAPACK
from Makefile.export, so it is no longer necessary to specify the
location of BLAS or LAPACK when configuring Tramonto.  To make this
work, it is not possible to use the --disable-libcheck option when
configuring Trilinos.  Due to this change, configure.ac will no longer
attempt to link to the AztecOO library, but rather will verify that
a header file from all required Trilinos packages can be found.