Added access of all continuation methods through Trilinos/LOCA. When configuring with  --with-noxloca, the code now uses NOX and LOCA from Trilinos instead of Tramonto's nonlinear solver and the old Loca. The input file has not changed.
The major change is the implementation of the phase transition
tracking method, which was done by creating a "PTGroup" and
"PTVector". Now, Tramonto presents the phase transition system
to the continuation package, where before the continuation
package owned the phase transition algorithm.

The new code appears to work for the problems studied. The
convergence histories are different since the backtracking method
in NOX is different then the update_solution method in Tramonto.
Also, the nonlinear convergence tolerances are different and the
continuation step-size control algoithms are different.

Two changes touched code that involves the "normal" code (that
does not use NOX and LOCA).

1) dft_BasicLinProbMgr.cpp:  I changed the logic so the matrix
is zeroed out even if it is the first fill. The problem is that
matrix entries are added even for residual-only fills. With NOX,
a residual-only fill precedes the first matrix fill, and so the
"new" matrix had non-zero entries in it.

2) Moved the prototype for "double calc_free_energy(FILE *, double **);"
down lower in the dft_continuation.h file.  If this prototype is above
#include <stdio.h> then it fails on the keyword FILE. If it is commented
out, as it was in the repository, then calc_free_energy() returns an
int instead of double. This is likely the reason that the CONT4 phase
transition example problem was failing for Laurie recently. This
is a "makeheaders" issue.