Committing implementation of new stencil package. Various properites of Each type of stencil (stencil radius, volume weights, quadrature definitions0 are now defined in stencil specific file.  Interaface routines found in (dft_switch_stencil.c) control the calls from the generic stencil calculation routines (still in dft_stencil.c) to the stencil specific routines.  New stencils based on delta and theta functions will require only addition of new stencil specific routines and small modifications of the interfaces.
In addition,two new dft_defs* routines have been added.  dft_defs_unknows.c contains definition of all of the
unknowns in the problem based on the functional switches set in the input file.  This was previously
found in dft_main.c.  This will be edited whenever a new functional is implemented.  The second
file dft_defs_stencil.c contains definitions for the Sten_Type[] array and a logical that is set
to TRUE for all of the Delta function based stencils.

The test harness was verified on 1 processor and 4 processors of the Mac Pro platform.