Hi Mike and Andy -
I am continuing to struggle with the math section.  After I wrote it last night, I had a look through other sections and the appendix and realized that the more general flavor served a purpose of allowing us to make the two physics specific appendices in a sensible way.  So, I've taken one more shot at putting this in more general terms - that I think are actually correct (within edits for indices)....

this new file is math_new.tex so adjust the main.tex file to compile this one.

--Laurie