This is the commit for the new Jain-Dominik-Chapman functionals.  Four new Examples have been added to the repository to test the code.  Some of the fill code has been rewritten into generic fill classes (see files dft_fill_CLSxxx). This had been done effectively for the mean field terms previously (when the routines were grouped in dft_utils.c), but we now have a general dft_fill_CLSchain for chain recursion relations (found in both CMS and in this new WJDC functional).  The new functional is invoked by setting Type_poly=3.
Work left to be done.  The current implementation has segment based density and mean
field paramters.  These need to be changed to component density and mean field parameters
as in the CMS functionals.  Also, the Schur solvers have not been implemented and tested
with this functional yet.  Both of these things should markedly improve the performance
of these functionals.
LJDF