These are my latest edits working towards component based density variables for WJDC functionals
This approach is available if you make edits to dft_fill_EL.c to turn on the
WJDC based cavity functions.  However, tests have not been promising at this point
as the field equations seem to make the entire matrix ill-conditioned when segment
densities are replaced with a combination of G_chain and field terms.

There are also some improvements in the numerical Jacobian capability and a
bug fix for polymer mixtures in dft_thermo_wjdc.c in this commit.

LF