Enabled Picard solver withing the Trilinos NOXLOCA solver. If the code is configured --with-noxloca, and Picard solver is requested (solver=2), then the NOX version of Picard will be called instead of the Tramonto version.
(The Newton code works the same way, with the configuration
choosing between the Tramonto and NOX solvers, not the input
file. Laurie talked about changing this in the near future.)

The NOX-Picard iteration works like this:
(1) NOX thinks it is doing regular Newton (but with an inaccurate
Jacobian).
(2) If Picard is requested, the Jacobian fill does nothing,
and the Jacobian-inverse operation is replaced with 100 iterations
of Tramonto's Picard iteration. (That means 10 Nonlinear Iterations
correspond to 1000 Picard iterations.)
(3) Convergence is checked only in NOX, which appears to be a much
stricter tolerance then the same settings mean in Tramonto.
(4) LOCA's zero-order continuation works: other types do not.


Small other change: Fixed a problem where a Nonlinear solve in
NOX would be followed by a continuation predictor step even if
no continuation was requested. This had resulted in overwritten
output files.