----------------------------------------------------------------------  Committing in .
 Modified Files:
 	src/dft_BasicLinProbMgr.cpp
 	src/dft_PolyA11_Epetra_Operator.cpp
 	src/dft_PolyA22_Epetra_Operator.cpp
 ----------------------------------------------------------------------

This commit fixes the Schur solver for cases where the primitive densities and
CMS-like variables differ in dimension.  The previous version assumed that
these dimensions were the same, which is now not true for some new formulations.

I have also added a fair amount of new sanity testing, to make sure that what
the user is passing in for data can actually be handled by the solver.  This
has exposed some incorrect input for some of the new functionals.   In particular,
values are being submitted that are in the upper triangle of the A11 block, which
goes against the block lower triangular form assumption of the Schur solvers.

I have run the code on POLY1_CMS_1D and it works properly.