Removed dependency of the executable (dft) on the library (libdft.a) in the Makefile.am.
The executable is linking against all the source.o files and not
the library, so there was no need for the dependency. When Jim and
I changed the executable to link agains the library instead of
the source.o file, then "make -j 4" continued to fail.
With this change, where the library and executable have no
interdependency, parallel "make -j 4" now seems to work.

I'm curious to know if this now works reliably for others as well.