This commit has added new Type_poly options.  They are as follows:
Type_poly=3 : WJDC - this is an implementation of the WJDC functionals
where both the field and the density variables are treated as segment based variables.
Note this option works with existing polymer Schur solver because the number of
variables found in the field and density equations are equal

Type_poly=4 : WJDC2 - this is an implementation of the WJDC functionals where the
field is treated as a component based variable while the density is treated as a
segment based variables.  This option does not currently work with the Schur solver
because the number of variables (field and density) is not equal in general.  Note
that this option uses the same physics scaling approach to handle the exp(mu) terms
as the WJDC option above.

Type_poly=5 : WJDC3 - this is a place holder for the a WJDC implementation where the
density and field variables are both treated on a component basis.  This is the optimal
implementation that should work with Schur solvers when complete. However, note that the
physics scaling routines will not be the same because we will not be able to modify density
equations on a per segment basis.

All three of these options will be available until we can establish that one implementation
is ideal.

--LJDF