commiting major changes to code interfaces for continuation capabilities.  The continuation code now has 4 pieces as follows.
1. dft_continuation.c contains only routines that interface Tramonto and LOCA.  Ths
fill will not be edited by a typical user.

2. dft_switch_continuation.c contains routines that switch among different continuation
types for initial values, parmeters, and output.  The set of continuation variables in
this file are ths set that have been implemented generally for a multicomponent or multisurface
problem.  These are the cases that are reflected in the dft_input.dat file, and this
file should not typically be edited by users or edited for special cases.

3.  dft_plugin_archived_continue.c contains all of the "special cases" that we as the
development team have implemented over the years.  It is a place for us to keep cases
that we use in publications or implement novel multiparameter continuations.

4.  dft_plugin_user_continue.c is a set of empty routines where a user can add their
specialized continuation algorithms without touching any real Tramonto code.

Previously all of these pieces were found in dft_continuation.c.  Note that all the
input files in the test harness are also updated here.

LF