Fixed bug in my implementation of the NOXLOCA_Tramonto_Vector.
This vector class was returning it's local length and not it's
global length. This was leading to different vector norms on
different processors within the NOX convergence tests, and we
were running into cases where some processors were finishing
and others returning for another Newton step. This was
causing the hangs with mismatched communications.

This also fixes test problem NLSOLVER4.