Temporary(?) fix for memory problem: The method "dft_BasicLinProbMgr::setupSolver()" is being called before every linear solve in both dft_newton and NOX solvers. There are some memory allocations that get repeated here.
I did a quick fix, similar to other methods in
this file, by adding a first line that is
  if (solver_ != Teuchos::null) return(0);
where solver_ is the pointer that was getting
re-allocated multiple times. All the tests pass.

I did not repeat this fix for other (Schur-based)
linear solvers. They appear to have the same problem
of a re-allocate; however, this function needs to
be called to get the correct results.

MIKE:: I think you need to sort this out. Is
solver_ getting repeatedly re-allocated in
setupSolver() for the Poly and HardSphere linear
solver? Valgrind was not catching the problem
in BasicLinProbMgr.


Separate change: in dft_linsolver_setup.c,
I had tried to set L_Schur in the input file to "0" to
force problems to use the Basic (global amtrix) solver,
and was finding this was being ignored. I added some
parentheses in the following line to make it behave as
I think it was supposed to.
<    else if (L_Schur && Type_poly == WJDC || Type_poly==WJDC3)   linsolver_setup_WJDCTYPE();
---
>    else if (L_Schur && (Type_poly == WJDC || Type_poly==WJDC3))   linsolver_setup_WJDCTYPE();

WARNING: If you have an input file with Type_Poly=5 (WJDC3)
and L_Schur=0, the code will use a different linear solver
then previously!