committing small change to printing and restart of WJDC_FIELD parameter.
This change removes the physics scaling parameter from the field parameter for printing
purposes, and then compensates on a restart by doing the multiplication by the scaling parameter.

This change is needed for proper restarts in cases where continuation in chemical
potential of a polymer chain is being done.  In thise case,
the scaling parameter cannot be easily updated as chemical potential is
varied. On restart, the scaling parameters are recomputed to be new and different
values, and so the WJDC_Field variable from the file was not correct.

--LF