This is the main commit associated with new code that separates the definition of uattr stencil from the details of the potential.
Other code modifications in this commit include :
(1) a bug fix for restarts when MF_EQ is present as a variable,
(2) some new code to calculate Rmin and Rzero for an arbitrary potential
(Rzero is the distance to the point where the cut and shifted potential of
interest will cross u=0).

Note that this code requires updates of input files.

Type_attr is redefined
with the following options:
#define NONE        -1
#define MFPAIR_RMIN_UMIN      0
#define MFPAIR_SIGMA_ZERO     1
#define MFPAIR_SIGMA_USIGMA   2
#define MFPAIR_SIGTOUMIN_UMIN 3
#define MFPAIR_RCSZERO_ZERO   4

Where type 0 is a WCA like definition of uattr with uattr=umin for r<rmin.
      type 1 has a uattr core value (from r=0 to r=sigma) of uattr=0.
      type 2 has a uattr core value (from r=0 to r=sigma) of uattr=u(sigma)
      type 3 has a uattr core value (from r=0 to r=sigma) of uattr=0, and
                 a uattr core value (from r=sigma to r=rmin) of uattr=min, and
      type 4 has a uattr core value (from r=0 to r=rzero) of uattr=0.

In addition Type_pairPot=LJ_SIGMATORCUT is now eliminated.  Just set Type_attr to 3 and Type_pairPot to
the usual LJ potential.

Finally, the MF_EQ option (where attractions are moved to the A12 part of the matrix is now set
by a parameter ATTInA22Block which is found in the nonlinear solver section along with the Physics_Scaling Switch.

10 tests were performed successfully, however final debugging will be completed based on data from the TH over
the next day or two.


LF