committing initial work on Tramonto GUI - to check it out, you will need to uncomment the call to dft_OptikaGUI() in dft_main.c.
This commit implements about 1/3 of the input file with many of the complex
dependencies in place.  The logic needed to select old dft_input.dat style files
or newer xml files is not yet implemented.  The file dft_GUI_toTramonto.cpp
begins the process of translating from the teuchos lists to the Tramonto
parameters.

more commits are coming in the next week.

LF