Fix to bug in dft_PolyA22_Epetra_Operator.
This changes the cmsOnCms matrix from a diagonal block (stored as a vector)
to a general sparse matrix (stored as a crsmatrix). Preconditioning is effected
by applying the inverse of the diagonal of the crsmatrix.

Warnings:
1) The PolyA22_Coulomb_Epetra_Operator needs further work. Currently, no test uses this
class.
2) This revision causes the SELF_ASSEMBLE3_FMT3_WCA_WJDC_1D test to hang (only in parallel) on my
machine. This failing test (really, hanging test) does not appear to be related to
this bug fix. Laurie and I are still looking into why this test is hanging.