Committing changes to Tramonto data input as part of work on GUI... The edits address several issues as listed below.
               /******* REFACTOR OF INPUT ROUTINES ************/
The previous input file (dft_input.c) had data read, data broadcast, and
some parameter manipulation all interlaced.  These functions have been separated.
Proc 0 reads are found in dft_input.c, parameter manipulations are in dft_setup_params.c, and
broadcasts are in dft_input_broadcast.c.

              /******* MULTIPLE OPTIONAL PATHS FOR DATA INPUT ************/
New logic in dft_setup_params.c, controls whether the static dft_input.dat file is read, or a GUI is opened,
or if data flows from the static file to the GUI.  Note that in order to read a static file AND open
the GUI it is required to add the string GUI to the top of the dft_input.dat file of interest.


              /****** RUNNING THE GUI WITH A PARALLEL JOB  *********/
This version allows for using the GUI when running  a parallel job as the GUI is only opened by Proc 0.
The other procs just wait until the GUI is closed.


              /******** GUI BUILD PROCESS *********/
Finally, the GUI build has now been redesigned based on compiler directives so the variable -DBUILD_GUI must be
defined specifically in the config script or the GUI will not be built.  This will allow for clean use of the
code on platforms that don't have QT.  Error checking against the build type has been implemented.

/********* TESTING ***************/
This code was tested based on both the Sanity and Checkin sets of TH exmple problems using the new ctest capability
on 8 processors of a Mac Pro.  Here is the summary output from the Sanity testing.

100% tests passed, 0 tests failed out of 33

Label Time Summary:
Checkin       = 225.00 sec
Continuous    = 225.00 sec
Nightly       = 225.00 sec
Sanity        = 225.00 sec

Total Test time (real) = 225.00 sec


At this time, I would strongly recommend that the new ctest capabilities be used prior to any new
source code checkins - they are really helpful :-)

LF