Changing number of iterations for this test by 0.57%.
Recent commits have changed the number of iterations incurred by this test by a
small percentage, resulting in test failure. (The physical quantities computed by
this test remain correct.) This commit changes the number of required iterations
by less than 1% to match with the number of incurred iterations with the current
code base.