finished work to separate file i/o from cwd.
If code is compiled without GUI, the paths for read and write file i/o are set deliberately to the cwd.
If the code is compiled with the GUI, there is an opportunity before setting up any physics to select directories.
In this case the input and output directories can be different.

LF