Fixed a bug in the TpetraHardSphereLinProbMgr- the preconditioner should be an inverse operator (type INVOP) that switches the apply and applyInverse methods.
Fixed a bug in the MultiVectorConverter. The arguments to the functions should
be passed by reference.

Implemented the option of doing the preconditioning and operator apply in half
the precision of the outer iterative solver for the TpetraPolyLinProbMgr and
the TpetraHardSphereLinProbMgr. This option can be turned on from the configure
file by specifying -DMIXED_PRECISION. This results in all the preconditioning
and operator applications being done in halfPrecision, where the linear solver
code is templated in scalar precision.