Committing two recent changes from Lisa:
1. New guess option (Iguess == 6), which produces a random guess for the densities.  Currently set up to work with diblock copolymers, so
this adds a small random number to one density and subtracts that same random number from the other density, at all points in space. There
is one new input parameter, related to the desired strength of the perturbation to the density.

2. The code now prints vtk format files of the density profiles if the Picard solver is used; a vtk file is saved with the same frequency
that the residual is printed to the screen.  The final converged result is also saved as a vtk file (as well as the normal Tramonto density
file).  These vtk files can be loaded into Paraview for easy visualization.

Both of these features are preliminary and need additional fine tuning.  All the test problems should still work. I am putting this in
now so Lisa and Mike Parks can work on things now.  I will fix/update and make things more general starting in August.

-Amalie