Fixing a bug in reflective boundaries with tethered chains.
Also, this commit comments out the diagonal test in dft_PolyA22_Epetra_Operator.cpp
at line 127 for diagonal only entries in the density on field part of the matrix.
This sub-block is no longer always diagonal in the case of tethered chains.
Note that preconditioners have not been adjusted to optimize solves for tethered chains.

LF