bug fix for semi-grand tethered systems.  Bug found in derivatives of field terms with respect to the tethered surface terms. numerical and analytical jacobians found to be exact using the POLYGRAFT3 test case.
Note that the if(debug_) check in the writeMatrix routines of both
dft_HardSphereLinProbMgr.cpp and dft_PolyLinProbMgr.cpp were disabled to
allow use of the numerical jacobian test functions in Tramonto.  A few
changes were made to the numerical jacobian routine in dft_newton.c to
make output more clear.

LF