Refining the selections for various test harness run categories. New time limits are imposed:
Sanity:  30 seconds
Checkin  100 seconds
Continuous 600 seconds
Nightly    2000 seconds
Occasional -- still working on this one.

The metric I used for inclusion of a case in any category was that the
1 processor timing had to be less than 1/2 of the time limits above
on my mac.

Also, as I've been checking these, I have found several cases where
the Schur solver fails on 1 processor.  All of these cases have been
changed to the general solver.  They are:

HS8
EXP
SW
DIFFUSION3
GEOM2, 6,32,33,34

One case was found where the Schur solver improved stability on 1 processor
as compared to the general solver.  It was

POLY7


We have looked at these cases previously and the thought was that there might
be somthing about the Nproc>1 case that stabilizes the Schur systems, but this
still might be an issue we need to consider.

LF